Local testing file is downloaded to frontend, but desktop portal is not prompted for file download dialogue
